+++ src/main/java/org/hdf5javalib/datasource/AbstractTypedStreamingSource.java
@@ -6,6 +6,7 @@ import org.hdf5javalib.file.dataobject.message.DataspaceMessage;
 import org.hdf5javalib.file.dataobject.message.DatatypeMessage;
 import org.hdf5javalib.file.dataobject.message.datatype.HdfDatatype;
 
+import java.math.BigInteger;
 import java.nio.ByteBuffer;
 import java.nio.ByteOrder;
 import java.nio.channels.FileChannel;
@@ -45,7 +46,7 @@ public abstract class AbstractTypedStreamingSource<T> {
                 .orElseThrow(() -> new IllegalStateException("DataspaceMessage not found"))
                 .getDimensions();
 
-        this.readsAvailable = Math.toIntExact(dimensions[0].getInstance(Long.class));
+        this.readsAvailable = dimensions[0].getInstance(BigInteger.class).intValue();
         this.datatype = headerPrefixV1.findMessageByType(DatatypeMessage.class)
                 .orElseThrow()
                 .getHdfDatatype();
@@ -53,7 +54,7 @@ public abstract class AbstractTypedStreamingSource<T> {
         if (dimensions.length == 1) {
             this.elementsPerRecord = 1;
         } else if (dimensions.length == 2) {
-            this.elementsPerRecord = dimensions[1].getInstance(Integer.class);
+            this.elementsPerRecord = dimensions[1].getInstance(BigInteger.class).intValue();
         } else {
             throw new IllegalArgumentException("Unsupported dimensionality: " + dimensions.length);
         }
