+++ src/main/java/org/hdf5javalib/file/dataobject/HdfObjectHeaderPrefixV1.java
@@ -109,7 +109,7 @@ public class HdfObjectHeaderPrefixV1 {
 
             // After writing 6 messages, jump to continuation offset if present
             if (i == 5 && optContinuationMessage.isPresent()) {
-                buffer.position(optContinuationMessage.get().getContinuationOffset().getInstance(Integer.class));
+                buffer.position(optContinuationMessage.get().getContinuationOffset().getInstance(BigInteger.class).intValue());
             }
         }
     }
